requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
discord.py
pydantic
httpx